requirements.txt
streamlit==1.52.2
firebase-admin==6.2.0
pandas==2.0.0
plotly==5.17.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
streamlit-autorefresh==1.0.1
streamlit_google_auth==1.1.8
scikit-learn==1.3.0
numpy==1.26.4
opencv-python==4.8.0.74
pyzbar==0.1.9